created a script to start the docker compose services and it only needs one parameter, the postgress passsword

+++ scripts/compose-up.sh
@@ -1,5 +1,14 @@
 #!/bin/bash
 
-docker compose down
+if [ "$#" -ne 1 ]; then
+    echo "Usage: $0 <POSTGRES_PASSWORD>"
+    exit 1
+fi
+
+POSTGRES_PASSWORD=$1
+
 export DOCKER_SOCK_GROUP_ID=$(stat -c '%g' /var/run/docker.sock)
+export POSTGRES_PASSWORD=$POSTGRES_PASSWORD
+
+docker compose down
 docker compose up -d
